docker/requirements.txt
GitPython
colorama
